+++ studio/src/app/utils/editor/parse-paragraphs.utils.tsx
@@ -2,7 +2,7 @@ import {h, JSX} from '@stencil/core';
 
 import {v4 as uuid} from 'uuid';
 
-import {Paragraph} from '@deckdeckgo/editor';
+import {Paragraph, convertStyle} from '@deckdeckgo/editor';
 
 import {ParseElementsUtils} from './parse-elements.utils';
 
@@ -27,8 +27,13 @@ export class ParseParagraphsUtils {
 
       const ParagraphElement: string = paragraph.data.nodeName;
 
+      const attributes: Record<string, string | number | boolean | undefined | Record<string, string>> = {
+        ...(paragraph.data.attributes || {}),
+        ...(paragraph.data.attributes?.style && {style: convertStyle(paragraph.data.attributes.style as string)})
+      };
+
       const result: JSX.IntrinsicElements = (
-        <ParagraphElement key={uuid()} paragraph_id={ignoreParagraphId ? undefined : paragraph.id}>
+        <ParagraphElement key={uuid()} paragraph_id={ignoreParagraphId ? undefined : paragraph.id} {...attributes}>
           {content}
         </ParagraphElement>
       );