+++ utils/deck/src/index.ts
@@ -1,2 +1,1 @@
-export * from './utils/parse-style.utils';
 export * from './utils/slides.utils';